Working build for x86

+++ include/string1.h
@@ -7,7 +7,7 @@ char* strtok1(char *str, const char* delim);
 char* trim1(char *str, char delim);
 uint32_t strlen1(char *str);
 uint32_t strcmp1(char *str1, char *str2);
-uint32_t strcpy1(char *str, char *buf);
+uintptr_t strcpy1(char *str, char *buf);
 int atoi1(char* str);
 
 #endif // STRTOK1_H_
