+++ src/parser.c
@@ -27,7 +27,7 @@ __USER_TEXT expr *expr_from_str(char *str) {
     /* Loop from 2nd element to next to last element, because
        we need to skip the " signs at start and end of the string */
     for(i = 1; i < size - 1; i++) {
-        expr *new_char = expr_new_val(CHAR, (uint32_t)str[i]);
+        expr *new_char = expr_new_val(CHAR, (uintptr_t)str[i]);
         expr *new_cons = expr_new_cons(new_char, NULL);
         if (first == NULL)
             first = new_cons;
@@ -45,7 +45,7 @@ __USER_TEXT int parse_quotation_symbol(token_t *tokens, int fd, char *name, expr
     int ret_code; expr *parsed;
     if ((ret_code = parse_tokens(tokens, fd, &parsed)) < 0)
         return ret_code;
-    *out = expr_new_cons(expr_new_val(SYMBOL, (uint32_t)name),
+    *out = expr_new_cons(expr_new_val(SYMBOL, (uintptr_t)name),
                          expr_new_cons(parsed, NULL));
     return 0;
 }
@@ -87,10 +87,10 @@ __USER_TEXT int parse_tokens(token_t *tokens, int fd, expr **out) {
         if ((ret_code = parse_quotation_symbol(tokens, fd, COMMA_AT_STR, out)) < 0)
             return ret_code;
     } else if (is_number(token)) {
-        *out = expr_new_val(NUMBER, (uint32_t)atoi1(token));
+        *out = expr_new_val(NUMBER, (uintptr_t)atoi1(token));
     } else if (is_boolean(token)) {
         int val = strcmp1(token, BOOL_STR_T) == 0 ? 1 : 0;
-        *out = expr_new_val(BOOLEAN, (uint32_t)val);
+        *out = expr_new_val(BOOLEAN, (uintptr_t)val);
     } else if (is_string(token)) {
         *out = expr_from_str(token);
     } else if (is_nil(token)) {
@@ -132,7 +132,7 @@ __USER_TEXT int parse_tokens(token_t *tokens, int fd, expr **out) {
         /* If not any of the above, then its a symbol */
         token_t symbol_name = my_malloc(strlen1(token) + 1);
         strcpy1(symbol_name, token);
-        *out = expr_new_val(SYMBOL, (uint32_t)symbol_name);
+        *out = expr_new_val(SYMBOL, (uintptr_t)symbol_name);
     }
 
     return 0;
